+++ R/createLocationID.R
@@ -61,6 +61,7 @@
 #' latitude <- c( 47.5, 0, 47.5, NA, 47.5)
 #'
 #' createLocationID(longitude, latitude)
+#' createLocationID(longitude, latitude, precision = 7)
 #' createLocationID(longitude, latitude, invalidID = "bad")
 #' createLocationID(longitude, latitude, algorithm = "digest")
 #'