+++ R/functionArgument_lintRules.R
@@ -9,6 +9,42 @@
 #' with parsing or formatting datetimes and helps check whether the appropriate
 #' timezone arguments are being explicitly used.
 #'
+#' \preformatted{
+#' timezoneLintRules <- list(
+#'   # base functions
+#'   "as.Date" = "tz",
+#'   "as.POSIXct" = "tz",
+#'   "as.POSIXlt" = "tz",
+#'   "ISOdate" = "tz",
+#'   "ISOdatetime" = "tz",
+#'   "strftime" = "tz",
+#'   "strptime" = "tz",
+#'   "Sys.Date" = "DEPRECATED", # Please don't use this function!
+#'   "Sys.time" = "DEPRECATED", # Please don't use this function!
+#'   # lubridate functions
+#'   "as_datetime" = "tz",
+#'   "date_decimal" = "tz",
+#'   "fast_strptime" = "tz",
+#'   "force_tz" = "tzone",
+#'   "force_tzs" = "tzone_out",
+#'   "interval" = "tzone",
+#'   "local_time" = "tz",
+#'   "make_datetime" = "tz",
+#'   "now" = "tzone",
+#'   "parse_date_time" = "tz",
+#'   "parse_date_time2" = "tz",
+#'   "today" = "tzone",
+#'   "with_tz" = "tzone",
+#'   "ymd" = "tz",
+#'   "ymd_h" = "tz",
+#'   "ymd_hm" = "tz",
+#'   "ymd_hms" = "tz",
+#'   # MazamaCoreUtils functions
+#'   "dateRange" = "timezone",
+#'   "timeRange" = "timezone",
+#'   "parseDatetime" = "timezone"
+#' )
+#' }
 timezoneLintRules <- list(
   # base functions
   "as.Date" = "tz",